Split ROS installation from Docker image
The requirements for this feature state that the install script
SHOULD be the same for both the Docker image and the direct installation
on a RaspberryPi 5. Therefore, a more 'barebone' base Docker image is
used, and the installation of the package dependencies was moved into
the `ros_entrypoint.sh`.

NOTES:

- The install script was tested inside a Docker container running the
  `Ubuntu 24.04 amd64` base image from Windows 11 amd64. The
  install script SHOULD be tested (and validated) on a RaspberryPi 5 with a
  fresh install of Ubuntu 24.04 arm64 BEFORE merging the PR.
- I tested the install script inside a Docker container running the
  `Ubuntu 24.04 arm64` bsae image from Windows 11 amd64, but I
  encountered the following issues:
  - The install script took more than 1 hour to run;
  - There were errors when trying to find certain packages or files.