Minor fixes to install script
This commit adds (temporary) Band-Aids to make the install script not crash:

- `rosdep init` error due to existing list file;
- hard-coded username for chown in `/` workaround;
- add `--cmake-clean-cache` to `colcon build` and links it to `source` to
  avoid skipping build;

The install script has been tested on an RPi5 8GB with an Ethernet
connection. The install script took ~20 minutes to complete. However, the
"test" ROS2 program (`racecar_teleop/teleop.launch.py`) is yet to be tested
on a racecar.

NOTES:

- The install script can easily take 30+ minutes to complete on a RPi5 4GB
with no heatsink/fan;
- The "temporary" Band-Aids SHOULD be replaced before A25 for a more elegant
  solution.