Add suggestions from chameau5050 to install script
This commits adds a function to get the RAM size of the device (later modi-
fications to the install script will adjust processing resources based on
the device's RAM size to ensure the installation works even on memory
contrained devices). This commit also adds arguments to `colcon build`
command.

NOTES:

- This commit does NOT make use of the added function in the install script.