Migrate ROS2 serial communication package to C++
Due to critical performance issues, caused by the additional overhead of
Python, of the serial communication package, several problems were
observed during A24. This commit migrates the Python package to a C++
package (that essentially translates the Python code into C++), which
SHOULD significantly reduce the overhead for serial communication
between the Rasperry Pi and the Arduino.

The migration seeks to eliminate the frequent loss of communication, and
heratic behavior observed in A24.

NOTES:

- The two packages used for serial communication come from other
  repositories, but adding these repositories as git submodules WOULD
  make maintaining this repository more difficult, so the `.git/`
  folder of the packages are NOT tracked.