Add "Band-Aid" for GRO830's lab issue with `blob_detector`
This commit adds a "Band-Aid" for (temporarily) fixing the issue with
the lab PCs' "bad" version of NumPy used when executing the
`blob_detector` script in GRO830's lab. This is the same issue as the
one with `labo_brushfire`, but because `blob_detector` is executed using
a launch file, using `python3` to execute the script is kind of a PITA.

The fix makes it so `ros2 launch` avoids the error.