Ensure required Gazebo resource paths are added to environment variable
This commit aims to fix the final issue with GRO830's lab
(exercise 12.4) by properly setting the GZ_SIM_RESOURCE_PATH environment
variable in the launch file for the Gazebo simulation.

The main culprit for this issue seems to be that the resource paths were
added to the incorrect environment variable. IGN_GAZEBO_RESOURCE_PATH is
for Gazebo Fortress (for ROS2 Humble). From now on, "GZ_SIM" is the
correct prefix to use for environment variables.

NOTES:

- This commit has NOT yet been fully tested on a lab PC, so do NOT merge
yet!