Replace hard-coded, mode dependent values for axis and button names
The motivation for this commit stems from the discrepancies between the
input modes of the Logitech F710 controller (DirectInput and XInput),
potentially causing confusion when trying to teleoperate the racecar.

Instead of directly reading the values from the `Joy` msg in the
callback, a local dictionary containing the values of the axes and the
buttons is updated before going through the logic of the callback. This
abstraction SHOULD make it so the same input in both input modes
result in the same command\*\* sent to the controller. Additionally,
this abstraction makes the code more readable (IMO).

> \*\* Since the triggers are treated as buttons in DirectInput mode,
> a command using the analog value of a trigger will not be equivalent
> in both modes.

The ROS parameter `input_mode` is added to the node to specify the input
mode. **IMPORTANT**: the input mode does NOT update automatically if the
input mode is changed on the F710 or if a new controller is used.

CHANGES:

- Add node argument `input_mode` to specify the input mode;
- Add (local) axes and buttons dictionaries;
- Add dictionary update function;
- Replace hard-coded `msg.<category>[<ID>]` with mode-independant
  values;
- Auto-format the `teleop` node file;
- Add mappings for the Xbox One Controller and Logitech F710 (XInput) in
  the documentation:
  - The DirectInput mappings WILL be added in an ulterior commit (I
    currently don't have a F710 to test the mappings);