Fix SLAM errors introduced in ab34974
The changes in ab34974 aimed to remove warnings when starting the Gazebo
simulation caused by a discrepancy in attributes between the `.gazebo`
file and the SDF model. However, commenting out these attributes made
the SLAM stopped working because it couldn't properly convert the LiDAR
messages, which in turn caused it so the map couldn't be fetched.

This commit rolls back the changes, and the SLAM now works properly.

NOTES:

- The attributes warnings have reappeared. Eventually, we will need to
address this issue, but for now it is considered low priority.