Fix `racecar_behaviors` issue with deadman switch
The issue was that the arbitration node didn't check the controller
inputs to verify the required buttons for controlling the racecar with
the behaviors were pressed, so it sent the command with the highest
priority whether or not the deadman switch and the
"autopilot-using-behaviors" were pressed, which would have been
problematic when testing the behaviors on the racecar (as one can
conclude with an autonomous vehicle ignoring the security measures).

The arbitration node now checks the `joy` topic for the required inputs,
and publishes only the "manual" commands if the requirements aren't met.