Use `joy.game_controller_node` for consistent gamepad mappings
`joy.joy_node` uses SDL's lower-level Joystick API: when used with the
Logitech F710, the indices in the `Joy` message differ between the
XInput mode and the DirectInput mode.

This commit aims to eliminate the confusion that can arise when the user
unintentionally switches the gamepad into XInput mode. The "joy" node
now uses SDL's GameController API, which offers consistent mappings in
the `Joy` message, independently of the mode the gamepad is in. The
indices of the buttons and axes in the logic of the `teleop` package
have been changed to keep the original physical mappings identical. All
launch files that include a "joy" node have been updated to use the
correct executable.

Additionally, the check for the minimum number of axes has been removed,
because the only gamepad used for controlling the racecar is the
Logitech F710, so it doesn't make sense to check for a potentially
incompatible gamepad.

NOTES:

- This commit has only been tested on a Xbox One controller (XInput
  mode).

TODO:

- Test GRO830's lab with a Logitech F710 in both XInput and DirectInput
  mode, and confirm the values obtained in the `Joy` message AND the
  `ctl_ref` message are identical for both modes BEFORE merging this PR.